Improve documentation for installing Windows on ARM
Improve the README.md. Two machines, one with Windows and one with
Ubuntu aren't needed, just the Windows machine to run the UUP dump
scripts and Rufus. With that, delete the scripts/uupdump-build-iso.sh
script since we'll run the commands in a Command Prompt.

Improve some of the grammar and wording of the steps to make them
clearer.